Refactor: Extract to Mixin::SelectAsteriskMethods
